pyproject.toml
[project]
name = "aismartfolder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.64.0",
    "pillow>=11.1.0",
    "python-dotenv>=1.0.1",
    "watchdog>=6.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.0.0",
    "pytest>=8.3.4",
]
